Decouple photo gallery screen
